bug fixes with video player
